+++ saleor/permission/utils.py
@@ -1,74 +1,11 @@
 from typing import TYPE_CHECKING, Iterable, Union
 
-from .auth_filters import (
-    AuthorizationFilters,
-    is_app,
-    is_staff_user,
-    is_user,
-    resolve_authorization_filter_fn,
-)
-from .enums import (
-    PERMISSIONS_ENUMS,
-    AccountPermissions,
-    AppPermission,
-    BasePermissionEnum,
-    ChannelPermissions,
-    CheckoutPermissions,
-    DiscountPermissions,
-    GiftcardPermissions,
-    MenuPermissions,
-    OrderPermissions,
-    PagePermissions,
-    PageTypePermissions,
-    PaymentPermissions,
-    PluginsPermissions,
-    ProductPermissions,
-    ProductTypePermissions,
-    ShippingPermissions,
-    SitePermissions,
-    get_permission_names,
-    get_permissions,
-    get_permissions_codename,
-    get_permissions_enum_dict,
-    get_permissions_enum_list,
-    get_permissions_from_codenames,
-    get_permissions_from_names,
-    split_permission_codename,
-)
-
-__all__ = [
-    "PERMISSIONS_ENUMS",
-    "is_app",
-    "is_staff_user",
-    "is_user",
-    "AppPermission",
-    "ChannelPermissions",
-    "CheckoutPermissions",
-    "DiscountPermissions",
-    "GiftcardPermissions",
-    "MenuPermissions",
-    "OrderPermissions",
-    "PagePermissions",
-    "PageTypePermissions",
-    "PaymentPermissions",
-    "PluginsPermissions",
-    "ProductPermissions",
-    "ProductTypePermissions",
-    "ShippingPermissions",
-    "SitePermissions",
-    "get_permission_names",
-    "get_permissions",
-    "get_permissions_codename",
-    "get_permissions_enum_dict",
-    "get_permissions_enum_list",
-    "get_permissions_from_codenames",
-    "get_permissions_from_names",
-    "split_permission_codename",
-]
+from .auth_filters import AuthorizationFilters, resolve_authorization_filter_fn
+from .enums import AccountPermissions, BasePermissionEnum
 
 if TYPE_CHECKING:
-    from ...account.models import User
-    from ...app.models import App
+    from ..account.models import User
+    from ..app.models import App
 
 
 def one_of_permissions_or_auth_filter_required(
@@ -115,7 +52,7 @@ def one_of_permissions_or_auth_filter_required(
 def permission_required(
     requestor: Union["User", "App", None], perms: Iterable[BasePermissionEnum]
 ) -> bool:
-    from ...account.models import User
+    from ..account.models import User
 
     if isinstance(requestor, User):
         return requestor.has_perms(perms)
